JSON+LD htl refactoring
