+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/product/Product.java
@@ -58,10 +58,6 @@ public interface Product extends Component, PageMetadata {
 
     String getSku();
 
-    Integer getReviewCount();
-
-    Double getReviewSummary();
-
     Price getPriceRange();
 
     Boolean getInStock();
