+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/product/package-info.java
@@ -13,7 +13,7 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-@Version("5.7.0")
+@Version("6.0.0")
 package com.adobe.cq.commerce.core.components.models.product;
 
 import org.osgi.annotation.versioning.Version;
