+++ bundles/core/src/test/java/com/adobe/cq/commerce/core/components/internal/models/v1/product/ProductImplTest.java
@@ -547,7 +547,6 @@ public class ProductImplTest {
         assertTrue(productModel.loadClientPrice());
         Page launch = context.pageManager().getPage("/content/launches/2020/09/14/mylaunch" + PAGE);
         Whitebox.setInternalState(productModel, "currentPage", launch);
-        System.out.println(productModel.loadClientPrice());
         assertFalse(productModel.loadClientPrice());
     }
